requirements.txt
mysql-connector
awscli

